Tidy build and dont wait

+++ build.sh
@@ -1,10 +1,13 @@
 #!/bin/bash
 
 export EXENAME=${1:-Launch}
-OUTPUTDIR="$(realpath ${2:build})"
+OUTPUTDIR="$(realpath ${2:-build})"
 
 dir=$(cd -P -- "$(dirname -- "$0")" && pwd -P)
 cd $dir
 
-rm -r bin
+if [ -d bin ]; then
+    rm -r bin
+fi
+
 dotnet build -c Release -r win10-x64 -o "$OUTPUTDIR"
